Add studio images tab
