+++ ui/v2.5/src/components/Studios/StudioDetails/StudioImagesPanel.tsx
@@ -1,0 +1,12 @@
+import React from "react";
+import * as GQL from "src/core/generated-graphql";
+import { studioFilterHook } from "src/core/studios";
+import { ImageList } from "src/components/Images/ImageList";
+
+interface IStudioImagesPanel {
+  studio: Partial<GQL.StudioDataFragment>;
+}
+
+export const StudioImagesPanel: React.FC<IStudioImagesPanel> = ({ studio }) => {
+  return <ImageList filterHook={studioFilterHook(studio)} />;
+};
